+++ tests/insertion.rs
@@ -799,9 +799,10 @@ struct Replaced;
 struct EntityVisibility;
 
 impl VisibilityFilter for EntityVisibility {
+    type ClientComponent = Self;
     type Scope = Entity;
 
-    fn is_visible(&self, client_component: Option<&Self>) -> bool {
+    fn is_visible(&self, client_component: Option<&Self::ClientComponent>) -> bool {
         client_component.is_some()
     }
 }
@@ -811,9 +812,10 @@ impl VisibilityFilter for EntityVisibility {
 struct ComponentVisibility;
 
 impl VisibilityFilter for ComponentVisibility {
+    type ClientComponent = Self;
     type Scope = ComponentScope<A>;
 
-    fn is_visible(&self, client_component: Option<&Self>) -> bool {
+    fn is_visible(&self, client_component: Option<&Self::ClientComponent>) -> bool {
         client_component.is_some()
     }
 }