+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - In `VisibilityFilter::is_visible` the second argument is now wrapped in an `Option`, and the arguments have been swapped: `self` is now the entity component, and the client component is the second argument.
+- `VisibilityFilter` now has a `ClientComponent` associated type, which allows using a different component for client entities. Set it to `Self` to preserve the old behavior.
 - `ClientStats` now represents only current client statistics. Use the `ConnectedClientStats` component for connected client statistics on the server.
 - `ConditionerConfig` now represents configuration only for a connected client on the server. Use the `GlobalConditionerConfig` resource for global server or client configuration.
 - `ServerPlugin::tick_schedule` now wrapped in an `Option`. You can set it to `None` to trigger replication by manually incrementing `ServerTick`.